MD_Product_SerialDao  - Modificado array columns para public static @#$ WS_Sync  - Modificado processamento do json md_product_serial_tracking adicionando HmAux com a pk de serial que ja teve seus trackings deletado, para que isso não seja feito mais de uma vez. Bug que estava acontecendo, como sempre que ele entrava no loop ele apagava todos os tracking do serial, caso o serial tivesse mais de um tracking somente o ultimo persistia, ja que a cada iteração do loop os outros eram deletados. @#$ Sql_Act020_002  - Corrigido erros na query  - Modificado query para pegar todos os campo da tabela md_product_serial e adicionando metodo que retorna todos os cmapos da tabela concatenados por # para usar na eração do HmAux @#$ Sync_Checklist_Sql_003  - Modificado query para na deleção, ignorar produtos que tenham a flag offline = 1 @#$ Act020  - Criado metodo closeDrawer para que o presenter feche o drawer após a busca local ser realziada  - Modificado esquema rodado caso não haja conexão na busca de serial, para primeiro fechar o drawer, na sequencia passa os valores para os metodos setRecordInfo() e loadProductSerialList() e somente depois verificar se a qtd de seriais retornados é > 0. Com essa mudança o resultado anteriormente buscado é limpo e caso nehum seja encontrado exibe msg de nenhum registro encontrado @#$ ToolBox_Inf  - Add comentario no metodo cleanOldSyncChecklistData(), informando que o metodo de fato só limpa o resgistro na tabela sync_checklist e que os form são deletados "naturalmente" no proximo sincronismo.

+++ app/src/main/java/com/namoadigital/prj001/service/WS_Sync.java
@@ -759,6 +759,9 @@ public class WS_Sync extends IntentService {
             //
             // Processamento Tracking do serial
             //
+            //HmAux que contem como chave a pk do serial ja deletados.
+            HMAux serialAlreadyDeleted = new HMAux();
+
             File[] files_serial_tracking = ToolBox_Inf.getListOfFiles_v2("md_product_serial_tracking-");
 
             for (File _file : files_serial_tracking) {
@@ -772,6 +775,9 @@ public class WS_Sync extends IntentService {
                 );
                 //
                 for (MD_Product_Serial_Tracking tracking : trackingList){
+                    String pk = tracking.getCustomer_code()+"."+
+                                tracking.getProduct_code()+"."+
+                                tracking.getSerial_code();
                     //Seleciona o serial do tracking para descobrir o serial_tmp
                     MD_Product_Serial dbSerial = serialDao.getByString(
                             new MD_Product_Serial_Sql_009(
@@ -782,13 +788,17 @@ public class WS_Sync extends IntentService {
                     );
                     //
                     if(dbSerial != null && dbSerial.getSerial_code() > 0){
-                        //Remove todos os trackings do serial
-                        trackingDao.remove(
-                                new MD_Product_Serial_Tracking_Sql_004(
-                                        tracking.getCustomer_code(),
-                                        tracking.getProduct_code(),
-                                        tracking.getSerial_code()
-                                ).toSqlQuery() );
+                        if(!serialAlreadyDeleted.containsValue(pk)) {
+                            //Remove todos os trackings do serial
+                            trackingDao.remove(
+                                    new MD_Product_Serial_Tracking_Sql_004(
+                                            tracking.getCustomer_code(),
+                                            tracking.getProduct_code(),
+                                            tracking.getSerial_code()
+                                    ).toSqlQuery());
+                            //
+                            serialAlreadyDeleted.put(MD_Product_Serial_TrackingDao.SERIAL_CODE,pk);
+                        }
                         //Atualiza serial_tmp no obj tracking e depois no banco
                         tracking.setSerial_tmp(dbSerial.getSerial_tmp());
                         //
